requirements.txt
spacy
fastcoref
neo4j
python-dotenv
fastapi
uvicorn
